@@ -11,46 +11,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.0",
-    "@angular/cdk": "^8.1.2",
-    "@angular/common": "^8.2.0",
-    "@angular/compiler": "^8.2.0",
-    "@angular/core": "^8.2.0",
-    "@angular/flex-layout": "^8.0.0-beta.26",
-    "@angular/forms": "^8.2.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/forms": "^8.2.14",
     "@angular/http": "^7.1.4",
-    "@angular/material": "~8.1.2",
-    "@angular/platform-browser": "^8.2.0",
-    "@angular/platform-browser-dynamic": "^8.2.0",
-    "@angular/router": "^8.2.0",
-    "core-js": "^3.1.4",
-    "rxjs": "^6.3.3",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.10.0",
-    "socket.io": "^2.2.0",
-    "@types/socket.io": "latest",
+    "@angular/material": "~8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@types/express": "latest",
-    "express": "latest"
+    "@types/socket.io": "latest",
+    "core-js": "^3.6.1",
+    "express": "latest",
+    "rxjs": "^6.5.4",
+    "socket.io": "^2.3.0",
+    "crypto-js": "latest",
+    "tslib": "^1.9.0",
+    "zone.js": "^0.10.2",
+    "body-parser": "latest"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.802.0",
-    "@angular/cli": "~8.2.0",
-    "@angular/compiler-cli": "^8.2.0",
-    "@angular/language-service": "^8.2.0",
-    "@types/jasmine": "^3.3.16",
-    "@types/jasminewd2": "^2.0.6",
-    "@types/node": "~12.6.8",
-    "codelyzer": "~5.1.0",
+    "@angular-devkit/build-angular": "^0.802.2",
+    "@angular/cli": "^8.2.2",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@types/jasmine": "^3.5.0",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^12.6.9",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.2.0",
     "karma-chrome-launcher": "~3.0.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
+    "karma-jasmine-html-reporter": "^1.5.1",
     "protractor": "~5.4.2",
     "ts-node": "~8.3.0",
     "tslint": "~5.18.0",
-    "typescript": "^3.4.0"
+    "typescript": "3.5.3"
   }
 }
