@@ -1,3 +1,7 @@
 export class User {
-  constructor(public name: string) {}
+  public name;
+  // constructor(public name: string) {
+    constructor(){
+    
+  }
 }
