@@ -1,3 +1,4 @@
 export const environment = {
-  production: true
+  production: true,
+  chatServerUrl: "www.yqblog.com:3000/#/chat"
 };
